+++ src/types/assert.ts
@@ -1,0 +1,4 @@
+export default function assert(condition: boolean, message: string): asserts condition {
+    if (!condition)
+        throw new Error(message);
+}